+++ Plugins/VMCLiveLink/Source/VMCLiveLink/VMCLiveLink.Build.cs
@@ -15,30 +15,35 @@ public class VMCLiveLink : ModuleRules
             "LiveLink",
             "OSC",
 
-            // JSON (FJsonValue, FJsonSerializer, FJsonObject etc.)
-            "Json",
+			// JSON (FJsonValue, FJsonSerializer, FJsonObject etc.)
+			"Json",
             "JsonUtilities",
 
-            // DeveloperSettings (UDeveloperSettings / project settings)
-            "DeveloperSettings",
+			// DeveloperSettings (UDeveloperSettings / project settings)
+			"DeveloperSettings",
 
-            // If you need LiveLink animation helpers / remap types
-            "LiveLinkAnimationCore"
+			// If you need LiveLink animation helpers / remap types
+			"LiveLinkAnimationCore"
         });
 
         PrivateDependencyModuleNames.AddRange(new[]
         {
-            // used by AutoDetectAndApplyMapping()
-            "AssetRegistry"
+			// used by AutoDetectAndApplyMapping()
+			"AssetRegistry"
         });
 
+        // Keep editor-only modules guarded so runtime packaging doesn't pull them in
         if (Target.bBuildEditor)
         {
             PrivateDependencyModuleNames.AddRange(new[]
             {
-                // Editor-only functionality (creating assets, factories, editor UI)
-                "UnrealEd",
-                "AssetTools"
+				// Editor-only functionality (creating assets, factories, editor UI)
+				"UnrealEd",
+                "AssetTools",
+
+                // Slate for editor UI implemented in runtime module under WITH_EDITOR
+                "Slate",
+                "SlateCore"
             });
         }
     }
