+++ Plugins/VMCLiveLink/Source/VMCLiveLinkEditor/VMCLiveLinkEditor.Build.cs
@@ -6,14 +6,30 @@ public class VMCLiveLinkEditor : ModuleRules
     {
         PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
 
-        PublicDependencyModuleNames.AddRange(new[]{
-            "Core","CoreUObject","Engine"
+        // Minimal public runtime dependencies for editor module consumers
+        PublicDependencyModuleNames.AddRange(new[] {
+            "Core",
+            "CoreUObject",
+            "Engine"
         });
 
-        PrivateDependencyModuleNames.AddRange(new[]{
-            "VMCLiveLink",          // runtime module
-            "LiveLink","LiveLinkInterface","LiveLinkEditor",
-            "Slate","SlateCore","UnrealEd","AssetTools"
+        PrivateDependencyModuleNames.AddRange(new[] {
+            "VMCLiveLink",          // runtime module (contains the UCLASS/header)
+			"LiveLink",
+            "LiveLinkInterface",
+            "LiveLinkEditor",
+
+			// Slate / editor UI
+			"Slate",
+            "SlateCore",
+            "EditorStyle",
+
+			// Editor-only functionality used by the factory UI and asset tooling
+			"UnrealEd",
+            "AssetTools"
         });
+
+        // Editor modules are only valid for editor builds; safe to rely on Target.bBuildEditor here,
+        // but this module is already an Editor module so the build system will only compile it for editor targets.
     }
 }
